+++ src/wallet/money/currencyUnit/builders/CurrencyUnitJSONStorageBuilder.java
@@ -1,4 +1,4 @@
-package wallet.money.currencyUnit;
+package wallet.money.currencyUnit.builders;
 
 import com.mashape.unirest.http.HttpResponse;
 import com.mashape.unirest.http.Unirest;
@@ -8,23 +8,18 @@ import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Objects;
 
-public class CurrencyUnitSQLStorageBuilder implements CurrencyUnitStorageBuilder{
+public class CurrencyUnitJSONStorageBuilder implements CurrencyUnitStorageBuilder {
 
-    private static CurrencyUnitSQLStorageBuilder instance;
-    private static String tableName;
-    private static Connection dbConnection;
+    private static CurrencyUnitJSONStorageBuilder instance;
 
     private static JSONArray currenciesWebJSONArray;
+    private static JSONArray currenciesJSONArray;
 
-    private CurrencyUnitSQLStorageBuilder() {
+    private CurrencyUnitJSONStorageBuilder() {
         JSONParser jsonParser = new JSONParser();
 
         String url = "https://www.nbrb.by/api/exrates/rates?periodicity=0";
@@ -36,6 +31,15 @@ public class CurrencyUnitSQLStorageBuilder implements CurrencyUnitStorageBuilder
         } catch (UnirestException | ParseException e) {
             e.printStackTrace();
         }
+
+        currenciesJSONArray = new JSONArray();
+    }
+
+    public static CurrencyUnitJSONStorageBuilder getInstance() {
+        if(instance == null) {
+            instance = new CurrencyUnitJSONStorageBuilder();
+        }
+        return instance;
     }
 
     public List<String> getBuildPlan() {
@@ -49,6 +53,11 @@ public class CurrencyUnitSQLStorageBuilder implements CurrencyUnitStorageBuilder
     }
 
     @Override
+    public void reset() {
+        currenciesJSONArray = new JSONArray();
+    }
+
+    @Override
     public void buildCurrencyUnit(String currencyString) {
         JSONObject currencyObject = null;
         for(Object object : currenciesWebJSONArray) {
@@ -59,38 +68,17 @@ public class CurrencyUnitSQLStorageBuilder implements CurrencyUnitStorageBuilder
 
         long currencyId = (long)currencyObject.get("Cur_ID");
         long currencyScale = (long)currencyObject.get("Cur_Scale");
-        try {
-            PreparedStatement preparedStatement = dbConnection.prepareStatement("INSERT INTO " + tableName +
-                    " (currencyId, currencyName, currencyScale) VALUES (?, ?, ?)");
-            preparedStatement.setLong(1,currencyId);
-            preparedStatement.setString(2,currencyString);
-            preparedStatement.setLong(3,currencyScale);
 
-            preparedStatement.executeUpdate();
+        JSONObject localCurrencyObject = new JSONObject();
+        localCurrencyObject.put("currencyName",currencyString);
+        localCurrencyObject.put("currencyId",currencyId);
+        localCurrencyObject.put("currencyScale",currencyScale);
 
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
+        currenciesJSONArray.add(localCurrencyObject);
     }
 
-    @Override
-    public void reset() {
-        try {
-            Statement statement = dbConnection.createStatement();
-            statement.execute("DROP TABLE " + tableName);
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-    }
-
-    public static CurrencyUnitSQLStorageBuilder getInstance(String tableName, Connection dbConnection) throws SQLException {
-        CurrencyUnitSQLStorageBuilder.tableName = tableName;
-        CurrencyUnitSQLStorageBuilder.dbConnection = dbConnection;
-
-        if(instance == null) {
-            instance = new CurrencyUnitSQLStorageBuilder();
-        }
-        return instance;
+    public JSONArray getResult() {
+        return currenciesJSONArray;
     }
 
 }
